+++ src/enums/NewBlockButtonStyle.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace benf\neo\enums;
+
+/**
+ * Pseudo-enum for representing new block button styles.
+ *
+ * @package benf\neo\enums
+ * @author Spicy Web <plugins@spicyweb.com.au>
+ * @since 3.6.0
+ */
+abstract class NewBlockButtonStyle
+{
+    public const Matrix = 'matrix';
+    public const Grid = 'grid';
+    public const List = 'list';
+}