+++ CHANGELOG.md
@@ -5,12 +5,12 @@
 ### Added
 - Added the `useNewBlockGrid` plugin setting
 - Added the `blockTypeIconSources` plugin setting
+- Added a plugin settings page
+- Added French translations (thanks @scandella)
 - Added `benf\neo\controllers\Configurator::actionRenderBlockType()`
 - Added `benf\neo\models\BlockType::$iconId`
 - Added `benf\neo\services\BlockTypes::EVENT_SET_CONDITION_ELEMENT_TYPES`
 - Added `benf\neo\services\BlockTypes::renderBlockTypeSettings()`
-- Added French translation (thanks @scandella)
-- Added a plugin settings page
 
 ### Deprecated
 - Deprecated `benf\neo\assets\SettingsAsset::EVENT_SET_CONDITION_ELEMENT_TYPES`; use `benf\neo\services\BlockTypes::EVENT_SET_CONDITION_ELEMENT_TYPES` instead